Renamed scripts properly with git commit.
taxIDermy: Added root colour to dictionary.

traverse_taxonomy: traversal stops just before taxid 1 and does not include it. This means that hits that have this node as the common node results in an error. Added an if loop a the end of the traversal to add root to the taxonomy array if $parent=1.